Исправление прав доступа, оптимизация вьюсетов
