+++ students/k33422/Plehanova_Daria/lr_3/mountaineering_club/mountaineering/permissions.py
@@ -1,7 +1,10 @@
-from rest_framework.permissions import IsAuthenticated
+from rest_framework.permissions import IsAuthenticated, SAFE_METHODS
 
 
 class CurrentUserOrAdmin(IsAuthenticated):
 	def has_object_permission(self, request, view, obj):
+		if request.method in SAFE_METHODS:
+			return True
+		
 		user = request.user
 		return user.is_staff or obj.user.pk == user.pk
